Create michal_frankel
This is a solution for HW 1  (following exercise): 

"Your friend is typing his name into a keyboard.  Sometimes, when typing a character c, the key might get long pressed, and the character will be typed 1 or more times. You examine the typed characters of the keyboard.  Return True if it is possible that it was your friend's name, with some characters (possibly none) being long pressed." (Solution)